refactor NavBar graphql to separate file
