+++ gatsby/src/graphql/NavBar.ts
@@ -1,0 +1,48 @@
+import { useStaticQuery, graphql } from "gatsby"
+
+export default getData() {
+  const data = useStaticQuery(
+    graphql`
+      query {
+        contentfulLandingPage {
+          title
+        }
+        allContentfulSection(
+          sort: {sort_id: ASC}
+          filter: {sort_id: {in: [5, 9, 10, 11, 12, 13]}}
+        ) {
+          nodes {
+            linkHref
+            linkText
+            title
+          }
+        }
+      }
+    `)
+
+  const title: string = data.contentfulLandingPage.title
+  const servicesHref: string = data.allContentfulSection.nodes[0].linkHref
+  const servicesText: string = data.allContentfulSection.nodes[0].title
+  const workHref: string = data.allContentfulSection.nodes[1].linkHref
+  const workText: string = data.allContentfulSection.nodes[1].linkText
+  const articlesHref: string = data.allContentfulSection.nodes[2].linkHref
+  const articlesText: string = data.allContentfulSection.nodes[2].linkText
+  const contactHref: string = data.allContentfulSection.nodes[3].linkHref
+  const contactText: string = data.allContentfulSection.nodes[3].linkText
+  const careersHref: string = data.allContentfulSection.nodes[4].linkHref
+  const careersText: string = data.allContentfulSection.nodes[4].linkText
+  const aboutHref: string = data.allContentfulSection.nodes[5].linkHref
+  const aboutText: string = data.allContentfulSection.nodes[5].linkText
+  const navMobile: string = ""
+
+  const items: NavItemInfo[] = []
+
+  items.push({ href: workHref, text: workText })
+  items.push({ href: servicesHref, text: servicesText })
+  items.push({ href: articlesHref, text: articlesText })
+  items.push({ href: contactHref, text: contactText })
+  items.push({ href: careersHref, text: careersText })
+  items.push({ href: aboutHref, text: aboutText })
+
+  return items
+}
